refactoring and packaging
