+++ emfrbm/compat.py
