+++ setup.py
@@ -48,11 +48,10 @@ setup(
     author_email='basaks@gmail.com',
     url='https://github.com/charlesmartin14/emf-rbm',
     packages=['.'],
-    package_dir={'emf-rbm': '.'},
+    package_dir={'emf-rbm': 'emfrbm'},
     include_package_data=True,
     entry_points={
-        'console_scripts': [
-        ]
+        'console_scripts': []
     },
     setup_requires=[NUMPY_VERSION],
     install_requires=[
@@ -93,7 +92,7 @@ setup(
         "Programming Language :: Python :: 3.4",
         "Programming Language :: Python :: 3.5",
         "Programming Language :: Python :: 3.6",
-        "Programming Language :: Python :: 3.7",
+        # "Programming Language :: Python :: 3.7",
         # add additional supported python versions
         "Intended Audience :: Science/Research",
         "Intended Audience :: Developers",
